qos: Disable irq notifier when qos request is removed
When a qos request is added with IRQ affinity, irq notifier for that irq
is added and without disabling the irq notifier if the irq is free it
throws a warning. So disable irq notifier when the qos request is
removed.

Change-Id: I50faa4ecbe1b632c0f0f203ca52faf18753c33d4
Signed-off-by: Srinivas Rao L <lsrao@codeaurora.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
[fix checkpatch issues]
Signed-off-by: celtare21 <celtare21@gmail.com>
(cherry picked from commit 0c0151004c3b0dcc869a26391b7d6b5e870f2496)